+++ indra/newview/llviewermenu.cpp
@@ -769,7 +769,7 @@ class LLAdvancedToggleRenderType : public view_listener_t
 		U32 render_type = render_type_from_string( userdata.asString() );
 		if ( render_type != 0 )
 		{
-			LLPipeline::toggleRenderTypeControl( reinterpret_cast<void*>(render_type) ); // <alchemy/>
+			LLPipeline::toggleRenderTypeControl( reinterpret_cast<void*>(static_cast<intptr_t>(render_type)) ); // <alchemy/>
 		}
 		return true;
 	}
@@ -785,7 +785,7 @@ class LLAdvancedCheckRenderType : public view_listener_t
 
 		if ( render_type != 0 )
 		{
-			new_value = LLPipeline::hasRenderTypeControl( reinterpret_cast<void*>(render_type) ); // <alchemy/>
+			new_value = LLPipeline::hasRenderTypeControl( reinterpret_cast<void*>(static_cast<intptr_t>(render_type)) ); // <alchemy/>
 		}
 
 		return new_value;
@@ -844,7 +844,7 @@ class LLAdvancedToggleFeature : public view_listener_t
 		U32 feature = feature_from_string( userdata.asString() );
 		if ( feature != 0 )
 		{
-			LLPipeline::toggleRenderDebugFeature( reinterpret_cast<void*>(feature) ); // <alchemy/>
+			LLPipeline::toggleRenderDebugFeature( reinterpret_cast<void*>(static_cast<intptr_t>(feature))); // <alchemy/>
 		}
 		return true;
 	}
@@ -859,7 +859,7 @@ class LLAdvancedCheckFeature : public view_listener_t
 
 	if ( feature != 0 )
 	{
-		new_value = LLPipeline::toggleRenderDebugFeatureControl( reinterpret_cast<void*>(feature) ); // <alchemy/>
+		new_value = LLPipeline::toggleRenderDebugFeatureControl( reinterpret_cast<void*>(static_cast<intptr_t>(feature))); // <alchemy/>
 	}
 
 	return new_value;
@@ -1075,7 +1075,7 @@ class LLAdvancedToggleInfoDisplay : public view_listener_t
 		
 		if ( info_display != 0 )
 		{
-			LLPipeline::toggleRenderDebug( reinterpret_cast<void*>(info_display) ); // <alchemy/>
+			LLPipeline::toggleRenderDebug( reinterpret_cast<void*>(static_cast<intptr_t>(info_display))); // <alchemy/>
 		}
 
 		return true;
@@ -1092,7 +1092,7 @@ class LLAdvancedCheckInfoDisplay : public view_listener_t
 
 		if ( info_display != 0 )
 		{
-			new_value = LLPipeline::toggleRenderDebugControl( reinterpret_cast<void*>(info_display) ); // <alchemy/>
+			new_value = LLPipeline::toggleRenderDebugControl( reinterpret_cast<void*>(static_cast<intptr_t>(info_display))); // <alchemy/>
 		}
 
 		return new_value;
